requirements.txt
git+https://github.com/Rapptz/discord.py
deta==1.0.1
discord-pretty-help==1.3.3